+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "svelte-app",
-    "version": "1.2.0",
+    "version": "1.3.0",
     "devDependencies": {
         "@babel/core": "^7.8.4",
         "@babel/preset-env": "^7.8.4",
