+++ src/js/view-helpers.js
@@ -29,3 +29,46 @@ export const pathDistance = path => {
 function toLatLng(coords) {
     return L.latLng(coords.latitude, coords.longitude);
 }
+
+const months = [
+    "Jan",
+    "Feb",
+    "Mar",
+    "Apr",
+    "May",
+    "Jun",
+    "Jul",
+    "Aug",
+    "Sep",
+    "Oct",
+    "Nov",
+    "Decr"
+];
+export const dateFormat = (date, fstr, utc) => {
+    utc = utc ? "getUTC" : "get";
+    return fstr.replace(/%[YmdHMS]/g, function(m) {
+        switch (m) {
+            case "%Y":
+                return date[utc + "FullYear"](); // no leading zeros required
+            case "%m":
+                m = 1 + date[utc + "Month"]();
+                break;
+            case "%d":
+                m = date[utc + "Date"]();
+                break;
+            case "%H":
+                m = date[utc + "Hours"]();
+                break;
+            case "%M":
+                m = date[utc + "Minutes"]();
+                break;
+            case "%S":
+                m = date[utc + "Seconds"]();
+                break;
+            default:
+                return m.slice(1); // unknown code, remove %
+        }
+        // add leading zero if required
+        return ("0" + m).slice(-2);
+    });
+};
